docker-compose/ml/requirements_lstm.txt
tensorflow-cpu
#tensorflow==2.9.1 
#seaborn
ipython 
nbformat 
numpy #==1.23.2 
scikit-learn #==1.1.2 
pandas #==1.4.3 
flask #==2.2.2 
matplotlib #==3.5.3
influxdb-client[ciso]